hipchat import: Disable PMs temporarily.
In the initial version of the program, I didn't
realize that Hipchat's export puts the same message
in multiple places.  There's some sense to what they
do--it makes each user's directory "complete" in a
sense--but it wasn't obvious that the same messages
were repeated.  (To give a concrete example, if user
1 sent a message to user 4, the same message would
appears in users/1/history.json and users/4/history.json.)

I'll fix this in a future commit, but there are some
related things to fix first that will make this easier
to put back.